requirements.txt
requests~=2.32.3
pytest~=8.3.3
allure-python-commons~=2.13.5